requirements.txt
flask
transformers
keybert
torch
newspaper3k
lxml_html_clean
gunicorn
